# openapi.yaml
openapi: 3.0.3
info:
  title: Narada LMS API
  description: |
    Comprehensive API for Narada LMS - A multilingual Learning Management System for Vedic education.
    
    **Architecture:** Modular backend with 6 domain modules:
    - Identity & Access: User management, authentication, authorization
    - Content Publishing: Tracks, chapters, text segmentation
    - Media Pipeline: Audio files, media segments, segment mappings
    - Batch & Cohort: Batches, enrollments, co-instructors, progress evaluation
    - Learning Delivery: Student-facing learning APIs, progress tracking
    - System Admin: Audit logs, system settings, monitoring
    
    **Multilingual Support:** Telugu (te), Hindi/Devanagari (hi), English/IAST (en)
    
    **Authentication:** JWT Bearer tokens. Obtain a token via ``POST /api/auth/login``. Protected endpoints require the ``Authorization: Bearer <token>`` header.
    
    **Authorization:** Role-based access control (student, instructor, content_manager, admin)
  version: 2.0.0
  contact:
    name: NaradaLMS Support
    email: support@narada-lms.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/tracks:
    get:
      tags: [Tracks]
      summary: List all tracks
      description: Get all published learning tracks (public access)
      security: []
      responses:
        '200':
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
    
    post:
      tags: [Tracks]
      summary: Create new track
      description: Create a new learning track - Content managers/Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackRequest'
      responses:
        '201':
          description: Track created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
  /api/tracks/{id}:
    get:
      tags: [Tracks]
      summary: Get track by ID
      description: Retrieve specific track details
      security: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Track details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
    
    put:
      tags: [Tracks]
      summary: Update track
      description: Update track details - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackRequest'
      responses:
        '200':
          description: Track updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
    
    delete:
      tags: [Tracks]
      summary: Delete track
      description: Delete track and all chapters - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Track deleted successfully
  /api/tracks/{id}/move:
    post:
      tags: [Tracks]
      summary: Reorder track
      description: Move track up or down - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [direction]
              properties:
                direction:
                  type: string
                  enum: [up, down]
      responses:
        '200':
          description: Track order updated

    
    post:
      tags: [Chapters]
      summary: Create new chapter
      description: Create a new chapter - Content managers/Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChapterRequest'
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
  /api/chapters/{chapterId}/details:
    get:
      tags: [Chapters]
      summary: Get chapter details
      description: Retrieve specific chapter
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      responses:
        '200':
          description: Chapter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
    
    patch:
      tags: [Chapters]
      summary: Update chapter
      description: Update chapter content or metadata - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChapterRequest'
      responses:
        '200':
          description: Chapter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
    
    delete:
      tags: [Chapters]
      summary: Delete chapter
      description: Delete chapter (cannot delete if status=published) - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      responses:
        '200':
          description: Chapter deleted successfully
        '400':
          description: Cannot delete published chapter
  /api/chapters/{chapterId}/status:
    patch:
      tags: [Chapters]
      summary: Publish/unpublish chapter
      description: Toggle chapter status (draft/published) - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [draft, published]
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
  /api/chapters/{chapterId}/move:
    post:
      tags: [Chapters]
      summary: Reorder or move chapter
      description: Move chapter up/down or to another track - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                direction:
                  type: string
                  enum: [up, down]
                toTrackId:
                  type: integer
      responses:
        '200':
          description: Chapter moved successfully


  # -----------------------------------------------------------------------------
  # NAMESPACED CONTENT STUDIO ROUTES (/api/content)
  # -----------------------------------------------------------------------------
  /api/content/tracks/{trackId}/chapters:
    get:
      tags: [Chapters]
      summary: List chapters for a track
      description: Get all chapters within a specific track - Namespaced
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chapter'
    
    post:
      tags: [Chapters]
      summary: Create chapter under track
      description: Create a new chapter directly within a track - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChapterRequest'
      responses:
        '201':
          description: Chapter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
  /api/content/chapters/{chapterId}:
    put:
      tags: [Chapters]
      summary: Update chapter
      description: Update chapter content or metadata - Namespaced
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChapterRequest'
      responses:
        '200':
          description: Chapter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
    
    delete:
      tags: [Chapters]
      summary: Delete chapter
      description: Delete chapter - Namespaced
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      responses:
        '200':
          description: Chapter deleted successfully
  /api/segments:
    get:
      tags: [Text Segments]
      summary: Get text segments for chapter
      description: Retrieve all segments for a chapter (filter by chapterId and optionally script)
      security: []
      parameters:
        - name: chapterId
          in: query
          required: true
          schema:
            type: integer
        - name: script
          in: query
          schema:
            type: string
            enum: [te, hi, en]
      responses:
        '200':
          description: List of text segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TextSegment'
  /api/segments/{segmentId}:
    patch:
      tags: [Text Segments]
      summary: Update text segment
      description: Update segment boundaries - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTextSegmentRequest'
      responses:
        '200':
          description: Segment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextSegment'
    
    delete:
      tags: [Text Segments]
      summary: Delete text segment
      description: Delete segment and all mappings - Content managers/Admin only
      parameters:
        - $ref: '#/components/parameters/SegmentId'
      responses:
        '200':
          description: Segment deleted successfully
  /api/content/chapters/{chapterId}/segments:
    get:
      tags: [Text Segments]
      summary: Get segments for chapter
      description: Retrieve all segments for a chapter with optional script filter - Namespaced
      parameters:
        - $ref: '#/components/parameters/ChapterId'
        - name: script
          in: query
          schema:
            type: string
            enum: [te, hi, en]
      responses:
        '200':
          description: List of text segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TextSegment'
    
    post:
      tags: [Text Segments]
      summary: Create text segment
      description: Create new text segment under chapter - Namespaced
      parameters:
        - $ref: '#/components/parameters/ChapterId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTextSegmentRequest'
      responses:
        '201':
          description: Segment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TextSegment'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /api/auth/login; send as Authorization Bearer <token>
  schemas:
...
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
        status:
          type: string
          enum: [draft, published]
        estimatedHours:
          type: integer
        createdBy:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  parameters:
...
    TrackId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...